requirements.txt
asgiref==3.3.1
certifi==2021.10.8
cffi==1.14.5
charset-normalizer==2.0.7
cryptography==3.4.6
Django==3.1.7
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.2
# djangorestframework-simplejwt==5.0.0
environ==1.0
idna==3.3
jwt==1.2.0
Pillow==8.3.2
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.0.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.7
